requirements.txt
Django			3.2.23
python-decouple		3.8
psycopg2    2.9.6
Pillow      8.4.0
xlwt			1.3.0
django-adminlte3	0.1.6
django-crispy-forms	1.13
django-filter   21.1
reportlab 3.6.6
xhtml2pdf		0.2.5

crispy-bootstrap4	2022.1